requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
frozenlist==1.5.0
idna==3.10
iso8601==2.1.0
multidict==6.1.0
obs-websocket-py==1.0
propcache==0.2.0
tomli==2.0.2
twitchio==2.10.0
typing_extensions==4.12.2
websocket-client==1.8.0
yarl==1.17.1
